requirements.txt
Flask>=2.0.1
Pillow>=9.0.0
google-api-python-client>=2.0.0
earthengine-api>=0.1.367
oauth2client>=4.1.3 